Refactor with Hono framework and clean HTML structure
- Migrated from raw fetch handler to Hono framework
- Split getHomePage() into logical sections: head, body, script
- Eliminated mega-template-literal syntax errors
- Used string concatenation for clean HTML generation
- Hono handles routing cleanly (/ for home, /room/:code for WS, /api/*, /story/:id)
- JavaScript now executes without 'Invalid or unexpected token' errors
- Fixed: console shows '✅ Game script loaded' - no syntax errors

Status: Framework migration complete
Next: Restore full game logic and test room creation flow
