+++ package.json
@@ -28,5 +28,8 @@
   },
   "engines": {
     "node": ">=18.0.0"
+  },
+  "dependencies": {
+    "hono": "^4.11.3"
   }
 }
